+++ boot/protocol/stiletto_video.h
@@ -3,8 +3,9 @@
 
 #include <stdint.h>
 
+
 typedef struct {
-    uint64_t *pFrame_buffer_base;
+    uint32_t *pFrame_buffer_base;
     uint64_t frame_buffer_sz;
     uint32_t horiz;
     uint32_t vert;
